requirements.txt
fs
torch
numpy
hydra-core
tensorflow
tensorboard
